Initial support for the 1.11 protocol.
