+++ src/Entities/ArrowEntity.cpp
@@ -167,7 +167,7 @@ void cArrowEntity::CollectedBy(cPlayer & a_Dest)
 			}
 		}
 
-		GetWorld()->BroadcastCollectEntity(*this, a_Dest);
+		GetWorld()->BroadcastCollectEntity(*this, a_Dest, 1);
 		GetWorld()->BroadcastSoundEffect("random.pop", GetPosX(), GetPosY(), GetPosZ(), 0.5, static_cast<float>(0.75 + (static_cast<float>((GetUniqueID() * 23) % 32)) / 64));
 		m_bIsCollected = true;
 	}
