+++ src/Entities/Pickup.cpp
@@ -234,7 +234,7 @@ bool cPickup::CollectedBy(cPlayer & a_Dest)
 		}
 
 		m_Item.m_ItemCount -= NumAdded;
-		m_World->BroadcastCollectEntity(*this, a_Dest);
+		m_World->BroadcastCollectEntity(*this, a_Dest, NumAdded);
 
 		// Also send the "pop" sound effect with a somewhat random pitch (fast-random using EntityID ;)
 		m_World->BroadcastSoundEffect("random.pop", GetPosX(), GetPosY(), GetPosZ(), 0.5, (0.75f + (static_cast<float>((GetUniqueID() * 23) % 32)) / 64));
